+++ test/test_helper.exs
@@ -1,2 +1,1 @@
 ExUnit.start()
-ForgeSdk.init(:forge_sdk)